add error emitting trait for better error handling and ergonomics
