+++ crates/ploke-tui/Cargo.toml
@@ -62,6 +62,9 @@ uuid = { workspace = true }
 
 futures = "0.3" # For .fuse() on streams
 
+# global state                                                                                  
+once_cell = "1.19"
+
 [dev-dependencies]
 tempfile = "3.4"
 wiremock = "0.5.14"
